\name{pa-package}
\alias{pa-package}
\docType{package}
\title{
  Brinson analysis for equity portfolios
}
\description{
  The pa package provides facilities for conducting brinson analysis for
  equity portfolios.
}
\details{
  \tabular{ll}{
    Package: \tab pa\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2011-12-16\cr
    Depends: \tab R (>= 2.3.0), methods, grid, ggplot2\cr
    License: \tab GPL (>= 2)\cr
    LazyLoad: \tab yes\cr
  }
    
}

\author{
  
  David Kane 
  Yang Lu <Yang.Lu@williams.edu>
  
  Maintainer: Yang Lu <Yang.Lu@williams.edu>
}
\keyword{ package }